LayoutTests/storage/domstorage/localstorage/simple-events.html

<html>
<head>
<script src="resources/clearLocalStorage.js"></script>
<script>

if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

function log(a)
{
    document.getElementById("logger").innerHTML += a + "<br>";
}

function finish()
{
    if (window.layoutTestController)
        layoutTestController.notifyDone()
}

function handleStorageEvent(e)
{
    log("Storage event fired:");
    log("Key           - " + e.key);
    log("New Value     - " + e.newValue);
    log("Old Value     - " + e.oldValue);
    if (e.uri)
        log("Event has a URI");
    log("");
}

function runTest()
{
    if (!window.localStorage) {
        log("window.localStorage DOES NOT exist");
        finish();
        return;
    }
    
    window.addEventListener("storage", handleStorageEvent, false);

    window.localStorage.setItem("FOO", "BAR");
    window.localStorage.setItem("FU", "BAR");
    window.localStorage.removeItem("FOO");
    window.localStorage.removeItem("FU");
    
    finish();
}

</script>
</head>
<body onload="runTest();">
This is a test to make sure localStorage mutations fire StorageEvents<br>
<div id="logger"></div>
</body>
</html>
