LayoutTests/storage/domstorage/localstorage/simple-events-expected.txt

This is a test to make sure localStorage mutations fire StorageEvents
Storage event fired:
Key - FOO
New Value - BAR
Old Value - null
Event has a URI

Storage event fired:
Key - FU
New Value - BAR
Old Value - null
Event has a URI

Storage event fired:
Key - FOO
New Value - null
Old Value - BAR
Event has a URI

Storage event fired:
Key - FU
New Value - null
Old Value - BAR
Event has a URI
